+++ src/main/java/org/mmarini/wheelly/apis/MapSector.java
@@ -27,55 +27,70 @@ package org.mmarini.wheelly.apis;
 
 import java.awt.geom.Point2D;
 
-import static java.lang.Math.*;
-import static org.mmarini.wheelly.apis.Utils.direction;
-import static org.mmarini.wheelly.apis.Utils.normalizeDegAngle;
-
 /**
  * MapSector keeps the presence of obstacles in the sector
  */
 public class MapSector {
-    public static final double MAX_SIGNAL_DISTANCE = 3;
+    /**
+     * Returns the contact sector
+     *
+     * @param location  the location
+     * @param timestamp the timestamp
+     */
+    public static MapSector contact(Point2D location, long timestamp) {
+        return new MapSector(location, MapSectorStatus.CONTACT, timestamp);
+    }
 
+    /**
+     * Returns the empty sector
+     *
+     * @param location  the location
+     * @param timestamp the timestamp
+     */
     public static MapSector empty(Point2D location, long timestamp) {
-        return new MapSector(location, timestamp, false);
+        return new MapSector(location, MapSectorStatus.EMPTY, timestamp);
     }
 
-    public static MapSector filled(Point2D location, long timestamp) {
-        return new MapSector(location, timestamp, true);
+    /**
+     * Returns the hindered sector
+     *
+     * @param location  the location
+     * @param timestamp the timestamp
+     */
+    static MapSector hindered(Point2D location, long timestamp) {
+        return new MapSector(location, MapSectorStatus.HINDERED, timestamp);
     }
 
-    public static MapSector unknown(Point2D location) {
-        return new MapSector(location, 0, false);
+    /**
+     * Returns the unknown sector
+     *
+     * @param location the location
+     */
+    static MapSector unknown(Point2D location) {
+        return new MapSector(location, MapSectorStatus.UNKNOWN, 0);
     }
 
     private final Point2D location;
     private final long timestamp;
-    private final boolean filled;
+    private final MapSectorStatus status;
 
-    /**
-     * Creates the MapSector
-     *
-     * @param location  the location of sector
-     * @param timestamp the timestamp of filled info (0 if unknown)
-     * @param filled    true if sector is filled
-     */
-    public MapSector(Point2D location, long timestamp, boolean filled) {
+    protected MapSector(Point2D location, MapSectorStatus status, long timestamp) {
         this.location = location;
+        this.status = status;
         this.timestamp = timestamp;
-        this.filled = filled;
     }
 
-    public MapSector clean(long timeout) {
-        return (timestamp <= timeout) ? setTimestamp(0) : this;
+    public MapSector clean(long validTimestamp) {
+        return !isUnknown() && getTimestamp() < validTimestamp
+                ? unknown() : this;
+    }
+
+    public MapSector contact(long timestamp) {
+        return MapSector.contact(getLocation(), timestamp);
     }
 
     public MapSector empty(long timestamp) {
-        return new MapSector(location, timestamp, false);
-    }
-
-    public MapSector filled(long timestamp) {
-        return new MapSector(location, timestamp, true);
+        return MapSector.empty(getLocation(), timestamp);
     }
 
     public Point2D getLocation() {
@@ -86,61 +101,35 @@ public class MapSector {
         return timestamp;
     }
 
-    public MapSector setTimestamp(long timestamp) {
-        return new MapSector(location, timestamp, filled);
+    public MapSector hindered(long timestamp) {
+        return MapSector.hindered(getLocation(), timestamp);
     }
 
-    public boolean hasObstacle() {
-        return isKnown() && filled;
+    public boolean isContact() {
+        return status.equals(MapSectorStatus.CONTACT);
     }
 
-    public boolean isKnown() {
-        return timestamp > 0;
+    public boolean isEmpty() {
+        return status.equals(MapSectorStatus.EMPTY);
     }
 
-    public MapSector setFilled(boolean filled) {
-        return new MapSector(location, timestamp, filled);
+    public boolean isHindered() {
+        return status.equals(MapSectorStatus.HINDERED);
+    }
+
+    public boolean isUnknown() {
+        return status.equals(MapSectorStatus.UNKNOWN);
     }
 
     public MapSector union(MapSector other) {
-        return other.timestamp > timestamp ? new MapSector(location, other.timestamp, other.filled) : this;
+        return other.timestamp > timestamp ? new MapSector(location, other.status, other.timestamp) : this;
     }
 
-    /**
-     * Returns the updated map sector.
-     * Updates the map sector with the result of a sensor signal.
-     * <p>
-     * The condition to update the status of a sector is that the distance of the signal is in range (> minDistance && < MAX_SIGNAL_DISTANCE)
-     * And the signal direction is in range of sector (direction within the direction of sector edge)
-     * And the sensor signal >= sector distance - receptive distance<br>
-     * The status of sector is set to empty if no echo signal is detected or the sensor distance > sector distance + sector size / 2 - threshold
-     * otherwise is set to filled
-     * <p>
-     *
-     * @param signal            the sensor signal
-     * @param receptiveDistance the receptive sector distance (distance from signal to set sector filled)
-     */
-    public MapSector update(RadarMap.SensorSignal signal, double minDistance, double receptiveDistance) {
-        double sectorDistance = signal.sensorLocation.distance(location);
-        boolean inRange = sectorDistance >= minDistance && sectorDistance <= MAX_SIGNAL_DISTANCE;
-        if (!inRange) {
-            return this;
-        }
-        double sectorDirection = direction(signal.sensorLocation, location);
-        double sectorDirFromSens = normalizeDegAngle(signal.sensorDirection - toDegrees(sectorDirection));
-        int a0 = (int) round(toDegrees(asin(receptiveDistance / sectorDistance)));
-        boolean inDirection = abs(sectorDirFromSens) <= a0;
-        if (!inDirection) {
-            return this;
-        }
-        if (!signal.isEcho()) {
-            return new MapSector(location,
-                    signal.timestamp,
-                    false);
-        }
-        return signal.distance >= sectorDistance - receptiveDistance
-                ? new MapSector(location, signal.timestamp,
-                signal.distance <= sectorDistance + receptiveDistance)
-                : this;
+    public MapSector unknown() {
+        return isUnknown() ? this : MapSector.unknown(getLocation());
+    }
+
+    private enum MapSectorStatus {
+        UNKNOWN, EMPTY, HINDERED, CONTACT
     }
 }